requirements.txt
streamlit
telethon
python-dotenv

